refactor(fasit): split command.go into per-noun files
Extract tenant, env, feature, rollout, deployment, and complete
helpers into separate files for maintainability. Fix errcheck
warnings on response body close and harden OAuth callback server
(localhost-only binding, read header timeout, nosec annotations).
